Altered Mutation and initialization.py
Mean Init now sets the Spread to zero to more accurately replicate Mean Init as it works in ORB. NormalInit now uses a halfnormal distribution to initialize the spread instead of a uniform one. Added sigma_spread and changed sigma to sigma_center in Normal Init to use for the distributions. Removed HalfnormalInit.
 Cleaned up the Normal Mutation. Halfnorm and Uniform Mutation were removed. Both Normal and Halfnormal Increase now also move the Center using random noise to better replicate the mutations done in OBR